+++ draft-packages/menu-list/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.2.7...@kaizen/draft-menu-list@1.2.8) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
+## [1.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.2.6...@kaizen/draft-menu-list@1.2.7) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
+## [1.2.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.2.5...@kaizen/draft-menu-list@1.2.6) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
+## [1.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.2.4...@kaizen/draft-menu-list@1.2.5) (2020-05-07)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
 ## [1.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.2.3...@kaizen/draft-menu-list@1.2.4) (2020-05-06)
 
 **Note:** Version bump only for package @kaizen/draft-menu-list