+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@2.0.5...@kaizen/draft-modal@2.0.6) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [2.0.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@2.0.4...@kaizen/draft-modal@2.0.5) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [2.0.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@2.0.3...@kaizen/draft-modal@2.0.4) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [2.0.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@2.0.2...@kaizen/draft-modal@2.0.3) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [2.0.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@2.0.1...@kaizen/draft-modal@2.0.2) (2020-05-07)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [2.0.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@2.0.0...@kaizen/draft-modal@2.0.1) (2020-05-06)
 
 **Note:** Version bump only for package @kaizen/draft-modal