+++ draft-packages/well/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.8...@kaizen/draft-well@2.4.9) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.4.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.7...@kaizen/draft-well@2.4.8) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.4.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.6...@kaizen/draft-well@2.4.7) (2020-05-08)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
+## [2.4.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.5...@kaizen/draft-well@2.4.6) (2020-05-07)
+
+**Note:** Version bump only for package @kaizen/draft-well
+
+
+
+
+
 ## [2.4.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-well@2.4.4...@kaizen/draft-well@2.4.5) (2020-05-06)
 
 **Note:** Version bump only for package @kaizen/draft-well